Refactors Asset, now ArtAsset2d.
Refactors Asset and updates references. Asset is now called ArtAsset2d
to better reflect the image based nature of the module. The 2d suffix is
unnecessary at this time, currently serving to future proof, in case the
decision is made to add support for WebGL.

Also updates Globals module to return a single object and define
dependencies on demand.