Reduces HTML requirements.
Engine can now be inserted into an html page with only a script tag.
Previously a canvas element would have also been required. Now the
Screen module checks for the canvas element and if it is unable to find
it inserts its own right at the top of the body.