Refactors Engine flow, Modules for scrolling.
Adds two new modules used for scrolling. The first is 'game/World' which
defines the maximum distance in the x and y coordinates of the game
world. The goal is to define the bounds of the camera, and any entities
in the game world, and to provide coordinates in the first quadrant.

The second module is 'render/Camera' which will be providing a platform
independent interface for rendering. The Camera follows an entity and
will use collision detection to determine what world objects need to be
rendered. The Camera will take care of translating the world coordinates
to screen coordinates.

'game/YageJS' and 'scripts/main' are refactored to reduce the complexity
of 'scripts/main'. Now all game specific code is handled by
'game/YageJS' in two functions, update() and draw(). 'scripts/main' will
handle calling those functions at 30 frames per second. Changing the
timing of either update() or draw() is trivial.