Updates Player to use new ArtAssetFactory.
'game/Player' now uses the 'render/ArtAssetFactory' module to load and
render it's art assets. This provides better resource management as each
asset is only loaded once, with references provided to various entities.

The 'render/ArtAssetFactory' module manages the loading of ArtAssets.
Each resources is only loaded once, with a new ArtAssetRef object passed
to each entity that requests a resource. Resources are loaded
dynamically when requested, provided they are not already in memory.

The ArtAssetRef object tracks the state of an object as used by the
entity that requests it. The reference also includes functions to update
the animation and fire the draw event.