Moves drawing to a queue system.
Camera now takes incoming draw events and places them in a queue for
processing rather than calling the artAssetFactory.draw() method as
events come in. This enables the camera to ensure that objects are drawn
at the right point in the cycle. After the screens been cleared and the
world is drawn the queue gets emptied and processed.