Adds JSON Scripts for managing world layout.
JSON scripts located in a games directory will provide for better
tooling of worlds. Inside each sub-directory in the games directory is
a JSON file sharing the same name. This file consists of a single array
listing all the worlds in the directory in the order they are to be
played through.