Adds rough collision detection.
World now cycles through every entity and checks to see if they collide
with another entity. This is done quickly using a bounding sphere that
encompasses each entity. If the bounding spheres collide the two
entities at fault are passed as arguments to the others collide()
method. There is where entities can make more complex collision checks
and respond to collision.

Only the owner of the collide() method should be modified. The entity
passed in is for reference purposes.